Change little things
* Add SleepTime prop to simulation
* Add a control to modify SleepTime
* Colorize Nutrients
* Change collecting nutrients -> set IsCollected to true
* add override to string stuff
* Change Energy datatype to float

+++ NoobsGameOfLife.Core/Information/CellInfo.cs
@@ -3,7 +3,7 @@
     public struct CellInfo
     {
         public Location Position { get; private set; }
-        public int Energy { get; private set; }
+        public double Energy { get; private set; }
 
         public CellInfo(Cell cell)
         {
